Update EJB timer QS to demonstrate both persistent and non-persistent timers.
