+++ ejb-timer/src/main/java/org/jboss/as/quickstarts/ejb/timer/ScheduleExample.java
@@ -1,0 +1,36 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2015, Red Hat, Inc. and/or its affiliates, and individual
+ * contributors by the @authors tag. See the copyright.txt in the
+ * distribution for a full listing of individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.as.quickstarts.ejb.timer;
+
+import javax.ejb.Schedule;
+import javax.ejb.Singleton;
+import javax.ejb.Timer;
+
+/**
+ * Demonstrates how to use the EJB's @Schedule.
+ *
+ * @author <a href="mailto:ozizka@redhat.com">Ondrej Zizka</a>
+ * @author Paul Ferraro
+ */
+@Singleton
+public class ScheduleExample {
+
+    @Schedule(second = "*/6", minute = "*", hour = "*", info = "ScheduleExample", persistent = true)
+    public void persistentTimer(Timer timer) {
+        TimeoutHandler.INSTANCE.accept(timer);
+    }
+}
